+++ packages/uni_app/lib/view/home/widgets/generic_home_card.dart
