--- packages/uni_app/lib/view/home/widgets/main_cards_list.dart
@@ -1,189 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import 'package:uni/generated/l10n.dart';
-import 'package:uni/model/providers/startup/session_provider.dart';
-import 'package:uni/utils/favorite_widget_type.dart';
-import 'package:uni/view/common_widgets/generic_card.dart';
-import 'package:uni/view/home/widgets/bus_stop_card.dart';
-import 'package:uni/view/home/widgets/exam_card.dart';
-import 'package:uni/view/home/widgets/exit_app_dialog.dart';
-import 'package:uni/view/home/widgets/restaurant_card.dart';
-import 'package:uni/view/home/widgets/schedule_card.dart';
-import 'package:uni/view/library/widgets/library_occupation_card.dart';
-
-typedef CardCreator = GenericCard Function(
-  Key key, {
-  required bool editingMode,
-  void Function()? onDelete,
-});
-
-class MainCardsList extends StatefulWidget {
-  const MainCardsList(
-    this.favoriteCardTypes, {
-    required this.saveFavoriteCards,
-    required this.isEditing,
-    required this.toggleEditing,
-    super.key,
-  });
-
-  final List<FavoriteWidgetType> favoriteCardTypes;
-  final void Function(List<FavoriteWidgetType>) saveFavoriteCards;
-  final bool isEditing;
-  final void Function() toggleEditing;
-
-  static Map<FavoriteWidgetType, CardCreator> cardCreators = {
-    FavoriteWidgetType.schedule: ScheduleCard.fromEditingInformation,
-    FavoriteWidgetType.exams: ExamCard.fromEditingInformation,
-    FavoriteWidgetType.busStops: BusStopCard.fromEditingInformation,
-    FavoriteWidgetType.restaurant: RestaurantCard.fromEditingInformation,
-    FavoriteWidgetType.libraryOccupation:
-        LibraryOccupationCard.fromEditingInformation,
-  };
-
-  @override
-  State<StatefulWidget> createState() {
-    return MainCardsListState();
-  }
-}
-
-class MainCardsListState extends State<MainCardsList> {
-  @override
-  Widget build(BuildContext context) {
-    // ignore: deprecated_member_use, see #1209
-    return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.surface,
-      body: BackButtonExitWrapper(
-        child: SizedBox(
-          height: MediaQuery.of(context).size.height,
-          child: widget.isEditing
-              ? ReorderableListView(
-                  onReorder: reorderCard,
-                  children: favoriteCardsFromTypes(
-                    widget.favoriteCardTypes,
-                    context,
-                  ),
-                )
-              : ListView(
-                  children: <Widget>[
-                    ...favoriteCardsFromTypes(
-                      widget.favoriteCardTypes,
-                      context,
-                    ),
-                  ],
-                ),
-        ),
-      ),
-      floatingActionButton:
-          widget.isEditing ? createActionButton(context) : null,
-    );
-  }
-
-  Widget createActionButton(BuildContext context) {
-    return FloatingActionButton(
-      backgroundColor: Theme.of(context).colorScheme.secondary,
-      onPressed: () => showDialog<void>(
-        context: context,
-        builder: (context) {
-          return AlertDialog(
-            title: Text(
-              S.of(context).widget_prompt,
-              style: Theme.of(context).textTheme.headlineSmall,
-            ),
-            content: SizedBox(
-              height: 200,
-              width: 100,
-              child: ListView(children: getCardAdders(context)),
-            ),
-            actions: [
-              TextButton(
-                child: Text(
-                  S.of(context).cancel,
-                  style: Theme.of(context).textTheme.bodyMedium,
-                ),
-                onPressed: () => Navigator.pop(context),
-              ),
-            ],
-          );
-        },
-      ), // Add FAB functionality here
-      tooltip: S.of(context).add_widget,
-      child: Icon(Icons.add, color: Theme.of(context).colorScheme.onPrimary),
-    );
-  }
-
-  List<Widget> getCardAdders(BuildContext context) {
-    final session = Provider.of<SessionProvider>(context, listen: false).state!;
-
-    final possibleCardAdditions = MainCardsList.cardCreators.entries
-        .where((e) => e.key.isVisible(session.faculties))
-        .where((e) => !widget.favoriteCardTypes.contains(e.key))
-        .map(
-          (e) => DecoratedBox(
-            decoration: const BoxDecoration(),
-            child: ListTile(
-              title: Text(
-                e
-                    .value(Key(e.key.index.toString()), editingMode: false)
-                    .getTitle(context),
-                textAlign: TextAlign.center,
-              ),
-              onTap: () {
-                addCardToFavorites(e.key, context);
-                Navigator.pop(context);
-              },
-            ),
-          ),
-        )
-        .toList();
-
-    return possibleCardAdditions.isEmpty
-        ? [Text(S.of(context).all_widgets_added)]
-        : possibleCardAdditions;
-  }
-
-  List<Widget> favoriteCardsFromTypes(
-    List<FavoriteWidgetType> cardTypes,
-    BuildContext context,
-  ) {
-    final userSession =
-        Provider.of<SessionProvider>(context, listen: false).state;
-    return cardTypes
-        .where((type) => type.isVisible(userSession?.faculties ?? []))
-        .where((type) => MainCardsList.cardCreators.containsKey(type))
-        .map((type) {
-      final i = cardTypes.indexOf(type);
-      return MainCardsList.cardCreators[type]!(
-        Key(i.toString()),
-        editingMode: widget.isEditing,
-        onDelete: () => removeCardIndexFromFavorites(i, context),
-      );
-    }).toList();
-  }
-
-  void reorderCard(
-    int oldIndex,
-    int newIndex,
-  ) {
-    final newFavorites =
-        List<FavoriteWidgetType>.from(widget.favoriteCardTypes);
-    final tmp = newFavorites[oldIndex];
-    newFavorites
-      ..removeAt(oldIndex)
-      ..insert(oldIndex < newIndex ? newIndex - 1 : newIndex, tmp);
-    widget.saveFavoriteCards(newFavorites);
-  }
-
-  void removeCardIndexFromFavorites(int i, BuildContext context) {
-    final favorites = List<FavoriteWidgetType>.from(widget.favoriteCardTypes)
-      ..removeAt(i);
-    widget.saveFavoriteCards(favorites);
-  }
-
-  void addCardToFavorites(FavoriteWidgetType type, BuildContext context) {
-    final favorites = List<FavoriteWidgetType>.from(widget.favoriteCardTypes);
-    if (!favorites.contains(type)) {
-      favorites.add(type);
-    }
-    widget.saveFavoriteCards(favorites);
-  }
-}
