+++ packages/uni_app/lib/view/home/widgets/restaurant_home_card.dart
@@ -8,7 +8,7 @@ import 'package:uni/model/entities/restaurant.dart';
 import 'package:uni/model/providers/lazy/restaurant_provider.dart';
 import 'package:uni/model/utils/day_of_week.dart';
 import 'package:uni/utils/navigation_items.dart';
-import 'package:uni/view/home/widgets2/generic_home_card.dart';
+import 'package:uni/view/home/widgets/generic_home_card.dart';
 import 'package:uni/view/lazy_consumer.dart';
 import 'package:uni/view/locale_notifier.dart';
 import 'package:uni/view/restaurant/widgets/restaurant_utils.dart';
